# Management-student/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GioiTinh": {
				"enum": [
					"Nam",
					"Nu",
					"Khac"
				],
				"type": "string"
			},
			"TrangThaiHoatDong": {
				"enum": [
					"Danghoc",
					"TamNgunghoc",
					"KhongConHoc"
				],
				"type": "string"
			},
			"VaiTroNguoiDung": {
				"enum": [
					"SinhVien",
					"GiangVien",
					"Admin"
				],
				"type": "string"
			},
			"CreateSinhVien": {
				"properties": {
					"MasoSinhVien": {
						"type": "string"
					},
					"HoVaTenSinhVien": {
						"type": "string"
					},
					"NgaySinh": {
						"type": "string",
						"format": "date-time"
					},
					"GioiTinhHocSinh": {
						"$ref": "#/components/schemas/GioiTinh"
					},
					"NoiSinh": {
						"type": "string"
					},
					"ChuyenNghanh": {
						"type": "string"
					},
					"KhoaHoc": {
						"type": "number",
						"format": "double"
					},
					"TrangThai": {
						"$ref": "#/components/schemas/TrangThaiHoatDong"
					},
					"VaiTro": {
						"$ref": "#/components/schemas/VaiTroNguoiDung"
					},
					"Email": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"SoLamDangNhapThatBai": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"KhongChoDangNhapToi": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"default": null
					}
				},
				"required": [
					"MasoSinhVien",
					"HoVaTenSinhVien",
					"NgaySinh",
					"GioiTinhHocSinh",
					"NoiSinh",
					"ChuyenNghanh",
					"KhoaHoc",
					"TrangThai",
					"Email",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSinhVien": {
				"properties": {
					"MasoSinhVien": {
						"type": "string"
					},
					"HoVaTenSinhVien": {
						"type": "string"
					},
					"NgaySinh": {
						"type": "string",
						"format": "date-time"
					},
					"GioiTinhHocSinh": {
						"$ref": "#/components/schemas/GioiTinh"
					},
					"NoiSinh": {
						"type": "string"
					},
					"ChuyenNghanh": {
						"type": "string"
					},
					"KhoaHoc": {
						"type": "number",
						"format": "double"
					},
					"TrangThai": {
						"$ref": "#/components/schemas/TrangThaiHoatDong"
					},
					"VaiTro": {
						"$ref": "#/components/schemas/VaiTroNguoiDung"
					},
					"Email": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"SoLamDangNhapThatBai": {
						"type": "number",
						"format": "double"
					},
					"KhongChoDangNhapToi": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"MasoSinhVien",
					"HoVaTenSinhVien",
					"NgaySinh",
					"GioiTinhHocSinh",
					"NoiSinh",
					"ChuyenNghanh",
					"KhoaHoc",
					"TrangThai",
					"Email",
					"UserName",
					"Password",
					"SoLamDangNhapThatBai",
					"KhongChoDangNhapToi"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TrangThaiHoatDongGiangVien": {
				"enum": [
					"DangDay",
					"TamNgungDay",
					"KhongConDay"
				],
				"type": "string"
			},
...
			"UpdateGiangVien": {
				"properties": {
					"MaSoGiangVien": {
						"type": "string"
					},
					"TenGiangVien": {
						"type": "string"
					},
					"NgaySinh": {
						"type": "string",
						"format": "date-time"
					},
					"NoiSinh": {
						"type": "string"
					},
					"GioiTinhGiangVien": {
						"$ref": "#/components/schemas/GioiTinh"
					},
					"ChuyenNghanh": {
						"type": "string"
					},
					"TrangThai": {
						"$ref": "#/components/schemas/TrangThaiHoatDongGiangVien"
					},
					"VaiTro": {
						"$ref": "#/components/schemas/VaiTroNguoiDung"
					},
					"Email": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"SoLamDangNhapThatBai": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"KhongChoDangNhapToi": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"default": null
					}
				},
				"required": [
					"MaSoGiangVien",
					"TenGiangVien",
					"NgaySinh",
					"NoiSinh",
					"GioiTinhGiangVien",
					"ChuyenNghanh",
					"TrangThai",
					"Email",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "management-student",
		"version": "0.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/API/v1/SinhVien/PostSinhVien": {
			"post": {
				"operationId": "PostSinhVien",
				"responses": {
					"201": {
						"description": "Insert sinh viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSinhVien"
							}
						}
					}
				}
			}
		},
		"/API/v1/SinhVien/GetOneSinhVien/{MasoSinhVien}": {
			"get": {
				"operationId": "GetOneSinhVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin sinh viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/SinhVien/GetAllSinhVien": {
			"get": {
				"operationId": "GetAllSinhVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin tất cả sinh viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/SinhVien/UpdateOneSinhVien/{MasoSinhVien}": {
			"put": {
				"operationId": "UpdateOneSinhVien",
				"responses": {
					"200": {
						"description": "Cập nhập thông tin sinh viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSinhVien"
							}
						}
					}
				}
			}
		},
		"/API/v1/SinhVien/DeleteOneSinhVien/{MasoSinhVien}": {
			"delete": {
				"operationId": "DeleteOneSinhVien",
				"responses": {
					"200": {
						"description": "Xoá sinh viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/API/v1/GiangVien/GetAllGiangVien": {
			"get": {
				"operationId": "GetAllSinhVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin tất cả giảng viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/GiangVien/UpdateOneGiangVien/{MaSoGiangVien}": {
			"put": {
				"operationId": "UpdateOneSinhVien",
				"responses": {
					"200": {
						"description": "Cập nhập thông tin giảng viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateGiangVien"
							}
						}
					}
				}
			}
		},
		"/API/v1/GiangVien/DeleteOneGiangVien/{MaSoGiangVien}": {
			"delete": {
				"operationId": "DeleteOneSinhVien",
				"responses": {
					"200": {
						"description": "Xoá giảng viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}